make the requested changes
The error you're seeing usually occurs when a plugin is already included in the classpath through another means, such as being part of the Android Gradle Plugin or another dependency.
For the androidx.navigation.safeargs and kotlin-parcelize plugins, it's possible that they are included as part of the Android Gradle Plugin or Kotlin plugin. To resolve this, you might not need to add them via the libs.versions.toml file. Instead, you can directly apply them in the build.gradle.kts file.